+++ packages/patternfly-4/gatsby-config.js
@@ -15,7 +15,7 @@ module.exports = {
       resolve: `gatsby-theme-patternfly-org`,
       options: {
         context: 'org',
-        hiddenPages: ['withOuia', 'Training'], // By title
+        hiddenPages: ['withOuia'], // By title
         showBanner: true, // PF3 banner
         sideNav: {
           core: [