+++ packages/patternfly-4/gatsby-theme-patternfly-org/components/example/example.js
@@ -51,6 +51,7 @@ export const Example = props => {
     const html = Prism.highlight(children, Prism.languages.javascript, 'javascript');
     return <pre dangerouslySetInnerHTML={{ __html: html }} />;
   }
+  const exampleName = title.replace(/-/g, ' ').replace(/  /g, '-');
   const fullscreenLink = `${location.pathname}/${slugger(title.toLowerCase())}`;
   const scope = useMDXScope();
   const codeBoxParams = getParameters(
@@ -93,11 +94,11 @@ export const Example = props => {
         darkMode ? ' pf-t-dark pf-m-opaque-200' : ''}${
         isFullscreen ? ' ws-preview-fullscreen' : ' ws-preview'}`} />
   );
-
+  
   return (
     <div className="ws-example">
       <AutoLinkHeader size="h4" headingLevel="h3" className="ws-example-heading">
-        {title.replace(/-/g, ' ').replace(/  /g, '-')}
+        {exampleName}
       </AutoLinkHeader>
       <LiveProvider
         scope={scope}
@@ -117,6 +118,7 @@ export const Example = props => {
                 className="ws-preview__thumbnail-link"
                 href={fullscreenLink}
                 target="_blank"
+                aria-label={`Open fullscreen ${exampleName} example`}
               >
                 <div style={previewContainerStyle}>
                   {Preview}
@@ -135,7 +137,8 @@ export const Example = props => {
           isFullscreen={isFullscreen}
           fullscreenLink={fullscreenLink}
           code={editorCode}
-          codeBoxParams={codeBoxParams} />
+          codeBoxParams={codeBoxParams} 
+          componentName={props.componentName}/>
         <LiveError />
       </LiveProvider>
     </div>