+++ packages/patternfly-4/gatsby-theme-patternfly-org/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.14](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.13...gatsby-theme-patternfly-org@1.2.14) (2019-12-19)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.13](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.12...gatsby-theme-patternfly-org@1.2.13) (2019-12-18)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.12](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.11...gatsby-theme-patternfly-org@1.2.12) (2019-12-17)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
+## [1.2.11](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.10...gatsby-theme-patternfly-org@1.2.11) (2019-12-17)
+
+**Note:** Version bump only for package gatsby-theme-patternfly-org
+
+
+
+
+
 ## [1.2.10](https://github.com/patternfly/patternfly-org/compare/gatsby-theme-patternfly-org@1.2.9...gatsby-theme-patternfly-org@1.2.10) (2019-12-09)